feat: Add Multiply Strings

+++ Math/Multiply Strings/java/Solution.java
@@ -2,7 +2,7 @@ class Solution {
     public String multiply(String num1, String num2) {
      int nums1= Integer.parseInt(num1);
      int nums2= Integer.parseInt(num2);
-     int ans=nums1*nums2;
-     return Integer.toString(ans);
+     long ans=nums1*nums2;
+     return Long.toString(ans);
          }
 }
